♻️ 読書記録を送信する処理をコンポーネント側に移動
